requirements.txt
azure-core==1.25.0
azure-storage-blob==12.13.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
Flask==2.2.2
idna==3.3
importlib-metadata==4.12.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
msrest==0.7.1
numpy==1.23.2
oauthlib==3.2.0
pandas==1.4.3
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
threadpoolctl==3.1.0
typing-extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.2
zipp==3.8.1
